# openapi.yaml
openapi: 3.0.3
info:
  title: MemStore API
  version: "1.0.0"
  description: >
    A minimal document store. Documents live in named collections and carry
    arbitrary JSON metadata. Listing supports opaque cursor pagination and a
    SQL-style metadata filter.
servers:
  - url: http://localhost:8000
security:
  - MemStoreKey: []
paths:
  /v1/collections:
    post:
      summary: Create a collection
      operationId: createCollection
      security:
        - MemStoreKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Collection created (or already existed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  created: { type: boolean }
        "401": { description: Missing or invalid X-MemStore-Key }
  /v1/collections/{collection}/documents:
    post:
      summary: Store a document
      operationId: storeDocument
      security:
        - MemStoreKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Document stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401": { description: Missing or invalid X-MemStore-Key }
        "404": { description: Collection not found }
    get:
      summary: List / paginate / filter documents
      operationId: listDocuments
      security:
        - MemStoreKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor.
          schema: { type: string }
        - name: filter
          in: query
          description: >
            SQL-style metadata filter, e.g. "category = 'invoice' AND amount > 1000".
            Operators: = != > < >= <=, joined with AND.
          schema: { type: string }
      responses:
        "200":
          description: A page of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items: { $ref: "#/components/schemas/Document" }
                  next_cursor:
                    type: string
                    nullable: true
                    description: Pass as cursor to fetch the next page; null when exhausted.
                  total: { type: integer }
        "401": { description: Missing or invalid X-MemStore-Key }
        "404": { description: Collection not found }
components:
  securitySchemes:
    MemStoreKey:
      type: apiKey
      in: header
      name: X-MemStore-Key
  schemas:
    Document:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        metadata:
          type: object
          additionalProperties: true